put ensure_unique_types on metadata, revert toml changes
