+++ codegen/Cargo.toml
@@ -13,24 +13,20 @@ homepage.workspace = true
 description = "Generate an API for interacting with a substrate node from FRAME metadata"
 
 [features]
-default = ["fetch_metadata"]
-fetch_metadata = ["dep:jsonrpsee", "dep:tokio", "dep:frame-metadata"]
+default = ["fetch-metadata"]
+fetch-metadata = ["dep:jsonrpsee", "dep:tokio"]
 web = ["getrandom/js"]
 
 [dependencies]
 codec = { package = "parity-scale-codec", workspace = true, features = ["derive"] }
-frame-metadata = { workspace = true, optional = true }
+frame-metadata = { workspace = true }
 heck = { workspace = true }
 proc-macro2 = { workspace = true }
 quote = { workspace = true }
 syn = { workspace = true }
 scale-info = { workspace = true }
 subxt-metadata = { workspace = true }
-jsonrpsee = { workspace = true, features = [
-    "async-client",
-    "client-ws-transport-native-tls",
-    "http-client",
-], optional = true }
+jsonrpsee = { workspace = true, features = ["async-client", "client-ws-transport-native-tls", "http-client"], optional = true }
 hex = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread"], optional = true }
 thiserror = { workspace = true }
