+++ codegen/src/lib.rs
@@ -50,15 +50,18 @@ mod error;
 mod ir;
 mod types;
 
-#[cfg(feature = "fetch_metadata")]
+#[cfg(feature = "fetch-metadata")]
 pub mod utils;
 
-#[cfg(all(feature = "web", feature = "fetch_metadata"))]
+#[cfg(all(feature = "web", feature = "fetch-metadata"))]
 compile_error!("subxt-codegen: the features 'web' and 'fetch_metadata' cannot be used together.");
 
 #[cfg(feature = "web")]
 #[allow(unused_imports)]
-pub use getrandom as _;
+const _: () = {
+    pub use getrandom as _;
+    use frame_metadata as _;
+};
 
 pub use self::{
     api::{GenerateRuntimeApi, RuntimeGenerator},
